+++ src/Retry.php
@@ -12,6 +12,8 @@ use Testo\Retry\Interceptor\RetryPolicyRunInterceptor;
  * Retry test on failure.
  *
  * A universal retry policy that can be applied to any test.
+ * When combined with {@see Repeat}, Retry wraps Repeat: each retry attempt
+ * runs the full repeat cycle. If any repetition fails, Retry decides whether to try again.
  *
  * @api
  */
@@ -19,15 +21,12 @@ use Testo\Retry\Interceptor\RetryPolicyRunInterceptor;
 #[FallbackInterceptor(RetryPolicyRunInterceptor::class)]
 final readonly class Retry implements Interceptable
 {
+    /**
+     * @param int<1, max> $maxAttempts Maximum number of attempts.
+     * @param bool $markFlaky Mark the test as flaky if it passed on retry.
+     */
     public function __construct(
-        /**
-         * Maximum number of attempts.
-         */
         public int $maxAttempts = 3,
-
-        /**
-         * Mark the test as flaky if it passed on retry.
-         */
         public bool $markFlaky = true,
     ) {}
 }
